Add capability for negative ampliude slepian waveforms
Sometimes it is desirable to have a negative amplitude slepian. Because of the definition of this waveforms, there's no way to set an 'amplitude' to a negative value (since the amplitude is found by properties of the qb spectrum). We have therefore added a simple functionality that just flips the amplitude of the waveform.