Fix CPMG Edge to Edge and Simultaneous Pulses
Fixed the CPMG module in "sequence_builtin.py" to correctly generate edge-to-edge pulses and simultaneous pulses.

NOTE: The edge-to-edge definition currently only works for Gaussian pulses.  To fix this for all pulses, we need to define the sequence duration in terms of the function "total_duration", which outputs the pulse widths for any type of pulse.