Modified CZ waveform
CZ waveform is now consistent with DiCarlo (arxiv: 1903.02492) paper. Couplings are scaled by a factor of two, the user is assumed to input coupling corresponding to |02> - |11> transition (typically: sqrt(2) * g). Also changed normalization condition for the waveform prefactors. The function theta(t) is now consistent with what is generated using the DiCarlo codebase.